mcp_server/pyproject.toml
[project]
name = "welfare-graph-mcp"
version = "0.1.0"
description = "MCP server for welfare-graph: 障害福祉知識グラフを Claude Desktop / Code から照会可能にする"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "welfare-graph contributors" },
]
keywords = ["mcp", "obsidian", "welfare", "disability-support", "knowledge-graph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "mcp>=1.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
neo4j = ["neo4j>=5.0"]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[project.scripts]
welfare-graph-mcp = "welfare_graph_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/welfare_graph_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"
